refactored test type
