+++ lib/Echidna.hs
@@ -21,6 +21,7 @@ import Echidna.Solidity
 import Echidna.Types.Campaign
 import Echidna.Types.Random
 import Echidna.Types.Signature
+import Echidna.Types.Test
 import Echidna.Types.Tx
 import Echidna.Types.World
 import Echidna.Processor
